Routing with Tanstack Router
We're building a single page app and we want to have multiple routes.
(isn't it funny how you can say a single page app but it has multiple
"pages", that we're calling routes? And you can see the page/route in
the address bar? Web development is a special kind of personality
disorder or something)

- Added the TanStack Router dependencies.
- Created new routes: __root, index, and order in the routes directory.
- Updated the components: the Order component became a route. And then the other components were updated to suit this.
- Added Links (that will be handled by the router) to link our "page" routes.

`__root.jsx` becomes essentially what the `App.jsx` was. `App` now
renders the Router component and the Router component will now render
the root route that is `__root.jsx`.

The router generates `routeTree.gen.ts` that contains all the information
about the routes and how they're mapped.

So, the router starts with `__root.jsx` and that's where the `Outlet` is.
We're on the `/` page/route, so the index route, `index.jsx` is rendered
into `Outlet`

Likewise, when we visit `/orders`, the router figures out what needs to be
rendered within/under the `Outlet` component.