GS Ch5 Change gets_free_shipping to better match the business domain
The business rule is to know if an order (cart) qualifies for
free shipping, not a total.

So, we calculate the *potential* order with the current item
in the cart by creating a new cart (remember `add_item`
is a calculation now) and testing it with the `get_free_shipping`.

This is what we did before when we made "planning" calculations.

This is neat.

- I've decided I'll make all the changes going forward in one TypeScript file rather than trying to keep multiple files by chapter. This is why I have git.

- I've added test file. Caught one mistake with calc total function.