GS Ch5 Improve add_item by extracing make_item from it
`add_item` was doing multiple things:
- it knew the structure of the cart and cart items.
- it was making items and updating the cart

Now we extract out the operation of making the item.
This means add_item does less.

One function knows how to work with carts, and the
other knows how to work with items. We can change/
evolve how Carts and Items are implemend independently
now!

For example if we wanted to change how Carts are
implemented, from an array to a hash map, we could
this with few changes