Implement the more efficient solution for Two Sum
This solution uses a map to track the numbers
we have already seen so that we can easily find
the number (complement) that completes our solution.

It reduces the complexity to O(n) but increases the
space from O(1) to O(n) because of the hash map.
(Hash Map look ups are generally O(n))