GS Ch8 Make freeTieClip a more straight forward implementation
We encapsulate the low level work (for loop) and implementation details (the Cart is a list) into a new function that we use to express all of the operations in
freeTieClip at the same level of detail.

In this case, we can see that `freeTieClip` is composed
(defined in terms of) of functions that operate on Carts
(and items). We're no longer doing any "primitive" operations on the Cart structure (looping over the array,
indexing into it).

`freeTieClip` now doesn't need to know HOW a
cart or items are implemented.