GS Ch8 Continue to refactor setPriceByName
We removed another direct arrow/use of the language layer by removing
the direct call to `.slice()`. `arraySet` will handle the copy-on-write
discipline for us.

However, we're still left indexing into the Cart array to get the Item
we want to change the price for.

The book stops here and will revisit it in a later chapter.
Perhaps we'll do a `getItemByName` by function?